+++ rust-backend/src/sniffer/mod.rs
@@ -203,8 +203,8 @@ impl Sniffer {
     }
 
     async fn handle_tcp_packet(&self, packet: TcpPacket<'_>) {
-        let source_port = packet.get_source();
-        let destination_port = packet.get_destination();
+        let source_port = packet.get_source() as i16;
+        let destination_port = packet.get_destination() as i16;
 
         let (port_pair, packet_direction): (PortPair, Option<domain::PacketDirection>) = {
             let ports = PORTS_TO_SNIFF.lock().await;
